Fixed generation of updated timelines & saving of images.
Now stores pictures inside the events and when saved stores them as well. Fixed when you click the ok button in update event it doesnt give nullpoint if no event is selected. Also fixed generation of updated timelines and error chacking if events is out of bounds. Also made opened timelines appear in the timeline listings.